+++ Assets/Charge Enemy.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 6013506944355332907}
   - component: {fileID: 4868008358126468717}
   - component: {fileID: 2657688402916316057}
-  - component: {fileID: 5559368271885973986}
   m_Layer: 0
   m_Name: Mesh
   m_TagString: Untagged
@@ -84,29 +83,6 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!136 &5559368271885973986
-CapsuleCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2226113839670669212}
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_LayerOverridePriority: 0
-  m_IsTrigger: 0
-  m_ProvidesContacts: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Radius: 0.5
-  m_Height: 2
-  m_Direction: 1
-  m_Center: {x: 0, y: 0, z: 0}
 --- !u!1 &6257359549676090643
 GameObject:
   m_ObjectHideFlags: 0
@@ -116,7 +92,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2488877618901478951}
-  - component: {fileID: 8344034285527924206}
+  - component: {fileID: 3465344172857142693}
+  - component: {fileID: 7579118032195940434}
   m_Layer: 0
   m_Name: Charge Enemy
   m_TagString: Enemy
@@ -140,7 +117,7 @@ Transform:
   - {fileID: 6013506944355332907}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &8344034285527924206
+--- !u!114 &3465344172857142693
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -149,6 +126,29 @@ MonoBehaviour:
   m_GameObject: {fileID: 6257359549676090643}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dec83457d04a4a4a9f0b3246cb08df14, type: 3}
+  m_Script: {fileID: 11500000, guid: 87882d5a155d1474c85fe60f1ae9049e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!136 &7579118032195940434
+CapsuleCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6257359549676090643}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 0
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Radius: 0.5
+  m_Height: 1
+  m_Direction: 1
+  m_Center: {x: 0, y: 0, z: 0}
