+++ Assets/Code/Character.cs
@@ -83,19 +83,3 @@ public class Player : Character
     }
 }
 
-
-//Parent class for all enemies, we can create children classes for each subtype
-public class Enemy : Character
-{
-    public Enemy(PosNode p) : base(p)
-    {
-	    throw new NotImplementedException();
-    }
-
-    public bool initateDeath()
-    {
-	    this.GetComponent<Renderer>().material.color = new Color(0, 0, 0);
-	    return true;
-    }
-}
-
