+++ Assets/Code/ChargeEnemy.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ChargeEnemy : Enemy
+{
+    public ChargeEnemy(PosNode p) : base(p)
+    {
+    }
+}
