+++ Assets/Untitled Scene.unity
@@ -279,6 +279,51 @@ PrefabInstance:
       insertIndex: -1
       addedObject: {fileID: 0}
   m_SourcePrefab: {fileID: 100100000, guid: 2fc0a15a2c3c8134f98e751b63b97abb, type: 3}
+--- !u!1 &311133080
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 311133082}
+  - component: {fileID: 311133081}
+  m_Layer: 0
+  m_Name: UnitController
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &311133081
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 311133080}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4a13cc171f2e48449814c5a660f0c38d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  movementSpeed: 1
+--- !u!4 &311133082
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 311133080}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &328699071
 GameObject:
   m_ObjectHideFlags: 0
@@ -638,22 +683,22 @@ PrefabInstance:
   serializedVersion: 2
   m_Modification:
     serializedVersion: 3
-    m_TransformParent: {fileID: 742086912}
+    m_TransformParent: {fileID: 0}
     m_Modifications:
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: -1.4764154
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -0.5201185
+      value: 0.5
       objectReference: {fileID: 0}
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: -0.38079047
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
@@ -695,50 +740,12 @@ PrefabInstance:
       propertyPath: m_Name
       value: Charge Enemy (1)
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 8344034285527924206, guid: 652dbe8f41caf174fa6b050decc41fd2, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 652dbe8f41caf174fa6b050decc41fd2, type: 3}
---- !u!4 &728845485 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
-    type: 3}
-  m_PrefabInstance: {fileID: 728845484}
-  m_PrefabAsset: {fileID: 0}
---- !u!1 &742086911
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 742086912}
-  m_Layer: 0
-  m_Name: Enemies
-  m_TagString: Enemy
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &742086912
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 742086911}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1.4764154, y: 1.0201185, z: 2.3807905}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children:
-  - {fileID: 3058455428084453655}
-  - {fileID: 728845485}
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!4 &788663272 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 287376318711958646, guid: 2fc0a15a2c3c8134f98e751b63b97abb,
@@ -1952,13 +1959,13 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1722671295}
   serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 1, z: -10}
+  m_LocalRotation: {x: 0.19720493, y: -0.6498734, z: 0.21390736, w: 0.70215267}
+  m_LocalPosition: {x: 6, y: 4, z: 1}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_LocalEulerAnglesHint: {x: 33.708, y: -84.651, z: 3.037}
 --- !u!114 &1722671300
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -2128,22 +2135,22 @@ PrefabInstance:
   serializedVersion: 2
   m_Modification:
     serializedVersion: 3
-    m_TransformParent: {fileID: 742086912}
+    m_TransformParent: {fileID: 0}
     m_Modifications:
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: 0
+      value: 1.4764154
       objectReference: {fileID: 0}
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -0.5201185
+      value: 0.5
       objectReference: {fileID: 0}
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: 0
+      value: 2.3807905
       objectReference: {fileID: 0}
     - target: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
         type: 3}
@@ -2185,17 +2192,12 @@ PrefabInstance:
       propertyPath: m_Name
       value: Charge Enemy
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 8344034285527924206, guid: 652dbe8f41caf174fa6b050decc41fd2, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 652dbe8f41caf174fa6b050decc41fd2, type: 3}
---- !u!4 &3058455428084453655 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 2488877618901478951, guid: 652dbe8f41caf174fa6b050decc41fd2,
-    type: 3}
-  m_PrefabInstance: {fileID: 3058455428084453654}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
@@ -2206,4 +2208,6 @@ SceneRoots:
   - {fileID: 328699072}
   - {fileID: 1026379996}
   - {fileID: 1289028433}
-  - {fileID: 742086912}
+  - {fileID: 3058455428084453654}
+  - {fileID: 728845484}
+  - {fileID: 311133082}
